kahoot_solver/requirements.txt
mss>=9.0.1
opencv-python>=4.9.0
numpy>=1.26.0
easyocr>=1.7.1
Pillow>=10.2.0
python-dotenv>=1.0.1
emergentintegrations
